requirements.txt
streamlit
pandas
pydeck
numpy
altair
requests
websocket-client
streamlit-autorefresh
colorama
